norm + pointer miss
