+++ srcs/common_utils.c
@@ -17,9 +17,9 @@ static int	search_path(char **abs_path, char **exec_path,
 			ft_free_str_arr(paths);
 			return (FAILURE);
 		}
-		if (stat(*abs_path, &ss) == 0 && (ss.st_mode))
+		if (stat(*abs_path, &ss) == 0 && (ss.st_mode & S_IXUSR))
 			break ;
-		if (stat(*abs_path, &ss) == 0 && exec_path == NULL)
+		if (stat(*abs_path, &ss) == 0 && *exec_path == NULL)
 			*exec_path = *abs_path;
 		else
 			free(*abs_path);
